% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoprintISB.R
\name{isbApp}
\alias{isbApp}
\title{A simple oncoprint App based on Institute for Systems Biology NCI Cloud Pilot}
\usage{
isbApp(bq)
}
\arguments{
\item{bq}{A \code{\link{src_bigquery}} object}
}
\description{
This function is a proof-of-concept for using the BigQuery interface
to TCGA data provided by the ISB Cancer Genomics Cloud Pilot. To use
this function, you must have a valid billing account with Google. If
you have an account with the ISB Cancer Genomics Cloud, then you can
use that account.
}
\examples{
\dontrun{
library(shiny)
library(dplyr)
isbApp(src_bigquery(project='isb-cgc',
   billing='isb-cgc-01-0006',dataset='tcga_201510_alpha')
}
}

